+++ CHANGES.md
@@ -1,4 +1,5 @@
 
+- Add `Bytes.Writer.writes_to_string`
 - `Bytesrw_sysrandom`: fix headers for musl libc (alpine.) (#7)
 
 v0.3.0 2025-11-04 Zagreb